:truck: Move ClayDataset and ClayDataModule into datamodule.py
Putting the DataLoader code in one file - datamodule.py. The regular torch Dataset classes are placed on top of the existing torchdata-based functions/classes.
