+++ trainer.py
@@ -17,10 +17,9 @@ from lightning.pytorch.cli import ArgsType, LightningCLI
 from lightning.pytorch.plugins.io import AsyncCheckpointIO
 
 from src.callbacks import LogIntermediatePredictions  # noqa: F401
-from src.datamodule import GeoTIFFDataPipeModule  # noqa: F401
+from src.datamodule import ClayDataModule, GeoTIFFDataPipeModule  # noqa: F401
 from src.model_clay import CLAYModule
 from src.model_vit import ViTLitModule  # noqa: F401
-from src.srm_datamodule import ClayDataModule
 
 
 # %%
